res okff.

+++ src/views/documents/DocumentCategoryView.js
@@ -16,6 +16,7 @@ import { useDispatch } from 'react-redux';
 import { documentView } from '../../features/documentSlice';
 import { toast } from 'react-toastify';
 import DocumentSubCategory from './DocumentSubCategory';
+import  Swal  from 'sweetalert2';
 
 function DocumentCategoryView() {
   
@@ -27,14 +28,31 @@ function DocumentCategoryView() {
   const [deleteDocument] = useDeleteDocumentMutation();
 
   const deleteHandel = async (id) => {
-    await deleteDocument(id);
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!',
+      width: 400,
+    }).then((result) => {
+      if (result.isConfirmed) {
+       deleteDocument(id);
+        Swal.fire(
+          'Deleted!',
+          'Your file has been deleted.',
+          'success'
+        )
+      }
+    })
   };
 
   if (isSuccess) {
     toast.success(data.message);
   }
 
-  console.log(data);
 
   return (
     <>
