+++ src/views/documents/DocumentSubSubCategoryView.js
@@ -17,6 +17,7 @@ import {
 import { useDispatch } from 'react-redux';
 import { documentView } from '../../features/documentSlice';
 import { toast } from 'react-toastify';
+import  Swal from 'sweetalert2';
 
 
 function DocumentSubSubCategoryView() {
@@ -29,7 +30,28 @@ function DocumentSubSubCategoryView() {
   const [deleteDocument] = useDeleteDocumentMutation();
 
   const deleteHandel = async (id) => {
-    await deleteDocument(id);
+    await 
+
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!',
+      width: 400,
+    }).then((result) => {
+      
+      if (result.isConfirmed) {
+        deleteDocument(id);
+        Swal.fire(
+          'Deleted!',
+          'Your file has been deleted.',
+          'success'
+        )
+      }
+    })
   };
 
   if (isSuccess) {
